requirements.txt
# Core data science
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.18.0

# Deep learning
torch>=1.9.0
torchvision>=0.10.0
tensorboard>=2.7.0

# Image processing
opencv-python>=4.5.0
openslide-python>=1.1.1
Pillow>=8.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
plotly-express>=0.4.1

# Dimensionality reduction
umap-learn>=0.5.0

# Utilities
tqdm>=4.62.0
joblib>=1.0.0
pyyaml>=5.4.0

# Bioinformatics
lifelines>=0.26.0
statsmodels>=0.13.0

# Optional: for advanced stain normalization
staintools>=2.1.2